requirements.txt
fastapi==0.115.12
sqlalchemy==2.0.40
pydantic==2.11.4
docker==7.1.0
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
jinja2==3.1.6
python-dateutil==2.9.0.post0